// src/types.ts
export interface TvShowMovie {
    backdrop_path: string;
    first_air_date?: string;
    id: number;
    name?: string;
    original_language: string;
    title?: string;
    release_date?: string;
    liked?: boolean;
    overview: string;
    popularity: number;
    vote_average: number;
    editMode?: boolean;
    writeComment?:string;
  }
  